+++ src/app/features/category-item-details/categories/container/add-category/add-category-container.component.ts
@@ -15,11 +15,11 @@ export class AddCategoryContainerComponent implements OnInit {
   // TODO Use a selector to offer popular category settings pre-filling the form.
   // See Issue #51.
   categoryForm = new FormGroup({
-    categoryName: new FormControl('cognitive_biases'),
-    label: new FormControl('Cognitive Biases'),
+    categoryName: new FormControl('fallacies'),
+    label: new FormControl('Fallacies'),
     language: new FormControl('en'),
     wdt: new FormControl('P31'),
-    wd: new FormControl('Q1127759')
+    wd: new FormControl('Q186150')
   });
   wikidataItemList: Category[];
   wikiListItems: Item[];
@@ -38,7 +38,7 @@ export class AddCategoryContainerComponent implements OnInit {
     });
     setTimeout(() => {
       this.default = 'Fallacies';
-      this.selectedPrefill = this.localCategories[1];
+      this.selectedPrefill = this.localCategories[0];
     }, 1600);
   }